fix: Classes definitions
