+++ src/Listener/Handler.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Lenra\App\Listener;
+
+abstract class Handler {
+    abstract public function handle(Request $request);
+}
