+++ src/Listener/Request.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Lenra\App\Listener;
+
+use Lenra\App\Api;
+
+class Request {
+    public function __construct(
+        public $props,
+        public $event,
+        public Api $api
+    ) {
+    }
+}
