+++ src/View/Renderer.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Lenra\App\View;
+
+abstract class Renderer {
+    abstract public function render(Request $request);
+}
