+++ src/View/Request.php
@@ -1,0 +1,13 @@
+<?php
+
+
+namespace Lenra\App\View;
+
+class Request {
+    public function __construct(
+        public $data,
+        public $props,
+        public $context
+    ) {
+    }
+}
